refactor(passcode): reorganize package structure and enhance navigation logic
- Reorganized files into specific subpackages: `biometricsSetup`, `components`, `di`, and `mifosPasscode`.
- Updated `MifosPasscodeNavigation` to use type-safe routes and added support for back navigation.
- Renamed `PasscodeModule` to `MifosAuthenticatorModule` and integrated it into the main Koin module configuration.
- Enhanced `MifosPasscode` composable with back press handling and improved callback naming for clarity.
- Added the passcode feature dependency to the navigation module.
