+++ feature/passcode/src/commonMain/kotlin/com/mifos/feature/passcode/biometricsSetup/BiometricsSetupScreenViewmodel.kt
@@ -7,7 +7,7 @@
  *
  * See https://github.com/openMF/mobile-wallet/blob/master/LICENSE.md
  */
-package com.mifos.feature.passcode
+package com.mifos.feature.passcode.biometricsSetup
 
 import androidx.lifecycle.viewModelScope
 import com.mifos.core.datastore.UserPreferencesRepository
